If a type is used as a superclass, force-generate its implicit initializers.
Our poor non-resilient vtables can't handle new methods being added to
superclasses, so when we do vtable layout we need to know that the
superclass is complete. If the superclass's initializers were never needed,
though, we would have never added them to the vtable, and we'd end up with
the wrong offset for all subclass methods.

<rdar://problem/17645615>

Swift SVN r19998