Fix <rdar://problem/17127126> swift compiler segfaults trying to generate a setter for a lazy property in a subclass of NSObject
We were setting 'final' on the storage for a lazy property, then calling "synthesizeTrivialSetter" to
synthesize the setter for the lazy property accessors.  This helper noticed that the storage was
final, and propagated the bit to the setter of the lazy property itself, causing chaos downstream.

While we're here, mark the storage of a lazy property as 'private' for good behavior.  Perhaps this
will allow special code for lazy property storage to be removed from PrintAsObjC and elsewhere.




Swift SVN r20220