Add a compile_swift_NAME target to avoid rebuilding .o files
Depending on the output .o file is not sufficient, since in a Makefile
build, the static and dynamic library builds will be in different
invocations of make, so we may build the .o twice in parallel if we
don't have the target dependency.  Not only is that wasteful, but swift
doesn't currently do this atomically, so we can get corrupt objects.

Swift SVN r20280