TryAddFinal - a walker that attempts to add 'final' when to decls.
TryAddFinal will recursively walk declarations (unless the class is
objc), attempting to add 'final' when appropriate. It can add final to
members that are not overridden, and are either private or internal
and in whole-compilation mode.

Currently does not add 'final' to classes, only properties and
methods. Should be invoked after normal type checking, so that we have
seen overridding definitions and established accessibility.

Swift SVN r20337