[ClangImporter] Make the 'value' field of plain C enums public again.
If an enum has a name but isn't declared with NS_ENUM or NS_OPTIONS, we
don't know how it's intended to be used, so we import it as a struct with
each enumerator as a value. The raw value of the enum used to be accessible
as the 'value' field, but that was mistakenly marked Private in the Great
Access Control Update.

We could consider making plain enums RawRepresentable (and just convert
to and from their raw type without checking anything), but that's something
to do later.

<rdar://problem/17753237>

Swift SVN r20355