Fix <rdar://problem/17755462> Predictable memory opts removes refcount operation
Don't remove otherwise dead allocations if there are uses that are the
destinations of copy_addr initializations that are not also takes of the
source. Doing so removes the retain of the source of the copy_addr.

There were no diffs building the stdlib -O0 with this change.

Swift SVN r20368