Add an 'isLValue' recursive type property.
As a first step toward addressing <rdar://problem/17507421>, introduce a type property we can use as a more general replacement for is<LValueType> that also catches tuples of @lvalues (and in the future, potential (@lvalue T)? optional-of-lvalue formations) and other types that need to be loaded in non-assignment contexts.

Swift SVN r20371