Add test case for '<rdar://problem/17796401> Swift Compiler never finishes compiling lazy collection reverse/filter/map'
This was a type checker performance issue, but testing that this
code also runs is also general goodness, as we need more of these
simple runtime tests to baseline functionality.

As part of this, consolidate the Fibonacci runtime test with
this test.  The new test file is generally named; we can
rename it if someone comes up with a better name.

Swift SVN r20559