Rip out my previous work that produced perplexing "inout writeback to computed property" errors when SILGen could determine that there was an inout writeback alias, and have the code instead perform CSE of the writebacks directly.
This means that we produce more efficient code, that a lot of things
now "just work" the way users would expect, and that the still erroneous
cases now get diagnosed with the "inout arguments are not allowed to 
alias each other" error, which people have a hope of understanding.

There is still more to do here in terms of detecting identical cases,
but that was true of the previous diagnostic as well.




Swift SVN r20658