IRGen: Don't read an ObjC isa field directly when building a generic subclass of a generic superclass.
A new generic class gets its metaclass's superclass pointer from the 
superclass's metaclass. On arm64 the superclass's metaclass may not be 
a pointer and must be fetched using object_getClass().

Three existing tests caught this bug but only when running on an iOS 8 
device. On iOS 7, the arclite implementation of objc_readClassPair() 
happened to rewrite the field before the runtime tripped over it.


Swift SVN r20811