Reduce depth of binary_operator_perf test.
I discussed this with JoeP, and he said that 100 additions is plenty to verify that we don't go exponential in this case. 1000 additions causes SILGen to exhaust the stack in release builds, but we don't have the time to refactor it to avoid the exhaustion right now. <rdar://problem/17898364>

Swift SVN r21035